+++ src/chart/rendering/markerrenderer.cpp
@@ -403,28 +403,6 @@ std::pair<Gfx::Color, Gfx::Color> MarkerRenderer::getColor(
 	auto finalBorderColor = actBorderColor * alpha;
 	auto itemColor = selectedColor * alpha * fillAlpha;
 
-	double highlight = 0.0;
-	double anyHighlight = 0.0;
-	auto markerInfo = plot->getMarkersInfo();
-	for (auto &info : markerInfo) {
-		auto allHighlight = 0.0;
-		info.second.visit(
-		    [&highlight,
-		        &allHighlight,
-		        idx = abstractMarker.marker.idx](int,
-		        const auto &info)
-		    {
-			    highlight += info.value.markerId == idx ? 1.0 : 0.0;
-			    if (info.value.markerId.has_value())
-				    allHighlight += info.weight;
-		    });
-		anyHighlight = std::max(anyHighlight, allHighlight);
-	}
-
-	auto highlightAlpha = 1 - (0.65 * anyHighlight) * (1 - highlight);
-	finalBorderColor = (finalBorderColor * highlightAlpha);
-	itemColor = (itemColor * highlightAlpha);
-
 	return std::make_pair(finalBorderColor, itemColor);
 }